backend/requirements.txt
# backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
langchain-community==0.2.19 # Changed to 0.2.19
langchain-core==0.2.43 # Changed to 0.2.19
langchain==0.2.17 
langgraph==0.1.7
python-dotenv==1.0.1
pydantic==2.8.2
sentence-transformers==2.7.0
faiss-cpu==1.8.0
unstructured[md]==0.18.9 # Updated to a version that supports Python 3.12
unstructured-inference==0.7.25
transformers==4.42.3
pandas
uvicorn[standard]
openpyxl # Good to have for Excel, though not strictly required for CSV
langchain-openai
langchain-huggingface

